+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [4.0.0-rc02] - XXXX-XX-XX
 
 ### Breaking
+- use solhint:recommended when no config is available, instead of exiting with
+  an error https://github.com/solhint-community/solhint-community/pull/135
 - exit with a different code when linter is configured incorrectly (255) vs
   when errors are found in linted files (1) https://github.com/solhint-community/solhint-community/pull/134
     - also exit eagerly when a misconfiguration is detected, to help the programmer